make entailment patterns mobile-friendly
following @bert-github's advice,
this PR changes the layout of the "entailment patterns" tables on narrow screens,
making all the cells stack vertically.

Note that, in this layout, the text present on the tables' first lines
("if S contains", "then S entails") is then repeated in each "row",
to keep them understandable.

Note also that, in fact, the text "then S entails" that is repeated in each "row"
is slightly less precise than the original table header
("then S RDF entails, recognizing D", then S RDFS entails, recognizing D").
But it should be clear from the context which entailment we are talking about.

Injecting exactly the right text in the small-screen layout is possible but
- it would increase the amount of text,
  which on small screens is not necessarily the best strategy, and
- it would require ad-hoc CSS rules for each tables,
  which makes it harder to maintain;
so I personally vote for keeping the more generic "then S entails" text.